+++ src/main/java/roomescape/controller/AdminController.java
@@ -1,51 +1,18 @@
 package roomescape.controller;
 
-import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.*;
-import roomescape.dto.ReservationResponse;
-import roomescape.dto.ReservationSaveRequest;
-import roomescape.service.ReservationService;
-
-import java.util.List;
+import org.springframework.web.bind.annotation.GetMapping;
 
 @Controller
 public class AdminController {
-    private final ReservationService reservationService;
-
-    public AdminController(ReservationService reservationService) {
-        this.reservationService = reservationService;
-    }
 
     @GetMapping("/admin")
-    public String adminPage() {
+    public String mainPage() {
         return "/admin/index";
     }
 
     @GetMapping("/reservation")
     public String reservationPage() {
         return "/admin/reservation-legacy";
-    }
-
-    @GetMapping("/reservations")
-    public ResponseEntity<List<ReservationResponse>> getReservations() {
-        var reservationResponses = reservationService.getReservations()
-                .stream()
-                .map(ReservationResponse::from)
-                .toList();
-        return ResponseEntity.ok(reservationResponses);
-    }
-
-    @PostMapping("/reservations")
-    public ResponseEntity<ReservationResponse> createReservation(@RequestBody ReservationSaveRequest request) {
-        var reservation = request.toReservation();
-        var savedReservation = reservationService.createReservation(reservation);
-        return ResponseEntity.ok(ReservationResponse.from(savedReservation));
-    }
-
-    @DeleteMapping("/reservations/{id}")
-    public ResponseEntity<Void> deleteReservation(@PathVariable Long id) {
-        reservationService.deleteReservation(id);
-        return ResponseEntity.ok().build();
     }
 }
