+++ src/main/java/roomescape/controller/ReservationController.java
@@ -1,0 +1,40 @@
+package roomescape.controller;
+
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+import roomescape.dto.ReservationResponse;
+import roomescape.dto.ReservationSaveRequest;
+import roomescape.service.ReservationService;
+
+import java.util.List;
+
+@RestController
+public class ReservationController {
+    private final ReservationService reservationService;
+
+    public ReservationController(ReservationService reservationService) {
+        this.reservationService = reservationService;
+    }
+
+    @GetMapping("/reservations")
+    public ResponseEntity<List<ReservationResponse>> getReservations() {
+        var reservationResponses = reservationService.getReservations()
+                .stream()
+                .map(ReservationResponse::from)
+                .toList();
+        return ResponseEntity.ok(reservationResponses);
+    }
+
+    @PostMapping("/reservations")
+    public ResponseEntity<ReservationResponse> createReservation(@RequestBody ReservationSaveRequest request) {
+        var reservation = request.toReservation();
+        var savedReservation = reservationService.createReservation(reservation);
+        return ResponseEntity.ok(ReservationResponse.from(savedReservation));
+    }
+
+    @DeleteMapping("/reservations/{id}")
+    public ResponseEntity<Void> deleteReservation(@PathVariable Long id) {
+        reservationService.deleteReservation(id);
+        return ResponseEntity.ok().build();
+    }
+}
