+++ src/test/java/roomescape/MissionStepTest.java
@@ -41,7 +41,7 @@ public class MissionStepTest {
     void 삼단계() {
         Map<String, String> params = new HashMap<>();
         params.put("name", "브라운");
-        params.put("date", "2023-08-05");
+        params.put("date", "2030-08-05");
         params.put("time", "15:40");
 
         RestAssured.given().log().all()
